% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEV.R
\name{NCEV}
\alias{NCEV}
\title{Incidence Matrix, Information Matrix, Canonical efficiency factor, Variance between associates and average variance}
\usage{
NCEV(design)
}
\arguments{
\item{design}{p-rep design in matrix form considering rows as blocks}
}
\value{
This function generates incidence matrix, information matrix, canonical efficiency factor, variance factor between associates and average variance factor for the input design
}
\description{
This function generates incidence matrix, information matrix, canonical efficiency factor,variance factor between associates and average variance for the input design
}
\examples{
library(pRepDesigns)
design=matrix(1:9, nrow = 3, ncol = 3)
NCEV(design)
}
\references{
Williams E, Piepho HP, Whitaker D (2011)<https://doi.org/10.1002/bimj.201000102>
}
