% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRep5.R
\name{pRep5}
\alias{pRep5}
\title{p-rep designs with unequal block sizes}
\usage{
pRep5(v, m, s)
}
\arguments{
\item{v}{Total number of treatments}

\item{m}{Positive integer (>=1)}

\item{s}{Positive integer (>=2)}
}
\value{
This function calculates design parameters (v, b, r, k), average
variance factors, and canonical efficiency factors of generated designs.
}
\description{
The primary purpose of this function is to generate various
proper p-rep designs for multi-environmental trials.
}
\examples{
\dontrun{
library(pRepDesigns)
pRep5(64, 2, 4)
}
}
